+++ migrations/2_deploy_diplocoin.js
@@ -1,0 +1,5 @@
+const DiploCoin = artifacts.require('DiploCoin');
+
+module.exports = async function (deployer) {
+  await deployer.deploy(DiploCoin);
+};