+++ src/app/models/contract.ts
@@ -1,5 +1,16 @@
 export class Contract {
     address: string;
+    name: string;
+    proposalThreshold: number;
+    quorumDenominator: number;
+    quorumNumerator: number;
+    timelock: string;
+    version: number;
+    votingDelay: number;
+    votingPeriod: number;
+    BALLOT_TYPEHASH: string;
+    COUNTING_MODE: string;
+    EXTENDED_BALLOT_TYPEHASH: string;
     quorum: number;
     minMinutes: number;
     sharesAddress: string;